+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Datasets downloaded with Clowder now include DataCite v4 XML files in the output /metadata folder for interoperability purposes.
 - Script to clean extractors' tmp files.
 - Script for RabbitMQ error queue cleanup.
+- Ability to use basic html formatting in the welcome message on the home page. [#51](https://github.com/clowder-framework/clowder/issues/51)
 
 ### Changed
 - Improved simple test to report all day success.